scraper/requirements.txt
# Core HTTP + parsing
httpx[http2]==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11

# Data + storage
pandas==2.2.3
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.36
supabase==2.28.3

# JavaScript rendering (used sparingly, only when required)
playwright==1.48.0

# Config + CLI
python-dotenv==1.0.1
click==8.1.7
pydantic==2.9.2

# Observability
structlog==24.4.0
tenacity==9.0.0

# Testing
pytest==8.3.3
